+++ lib/drawer_scaffold.dart
@@ -1,6 +1,5 @@
 import 'dart:math';
 
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
 import 'menu_screen.dart';
@@ -11,38 +10,6 @@ typedef Widget DrawerScaffoldBuilder(
 
 /// a Scaffold wrapper
 class DrawerScaffold extends StatefulWidget {
-  final List<SideDrawer> drawers;
-  @deprecated
-  final Screen contentView;
-  final ScreenBuilder builder;
-
-  final AppBar appBar;
-  final Direction mainDrawer;
-  final DrawerScaffoldController controller;
-  final double cornerRadius;
-  final bool extendedBody;
-  final bool enableGestures;
-  final Widget floatingActionButton;
-  final Widget bottomNavigationBar;
-  final FloatingActionButtonLocation floatingActionButtonLocation;
-  final FloatingActionButtonAnimator floatingActionButtonAnimator;
-  final List<BoxShadow> contentShadow;
-  final Widget bottomSheet;
-  final bool extendBodyBehindAppBar;
-  final List<Widget> persistentFooterButtons;
-  final bool primary;
-  final bool resizeToAvoidBottomInset;
-  final bool resizeToAvoidBottomPadding;
-
-  /// Listen to offset value on slide event for which [SideDrawer]
-  final Function(SideDrawer, double) onSlide;
-
-  /// Listen to which [SideDrawer] is opened (offset=1)
-  final Function(SideDrawer) onOpened;
-
-  /// Listen to which [SideDrawer] is closed (offset=0)
-  final Function(SideDrawer) onClosed;
-
   DrawerScaffold({
     this.appBar,
     this.contentShadow = const [
@@ -55,7 +22,6 @@ class DrawerScaffold extends StatefulWidget {
     ],
     this.drawers,
     this.cornerRadius = 16.0,
-    this.contentView,
     this.controller,
     this.extendedBody = false,
     this.bottomNavigationBar,
@@ -64,29 +30,70 @@ class DrawerScaffold extends StatefulWidget {
     this.floatingActionButtonAnimator,
     this.builder,
     this.enableGestures = true,
-    this.mainDrawer = Direction.left,
-    Key key,
+    this.defaultDirection = Direction.left,
+    Key? key,
     this.bottomSheet,
     this.extendBodyBehindAppBar = false,
     this.persistentFooterButtons,
     this.primary = true,
     this.resizeToAvoidBottomInset,
-    this.resizeToAvoidBottomPadding,
     this.onSlide,
     this.onOpened,
     this.onClosed,
+    this.backgroundColor,
   }) : super(key: key);
+
+  /// List of drawers
+  final List<SideDrawer>? drawers;
+
+  /// Screen Builder => Widget ScreenBuilder<T>(BuildContext context, T? id)
+  final ScreenBuilder? builder;
+
+  /// Appbar for Scaffold
+  final AppBar? appBar;
+
+  /// set default drawer from which [Direction], default: [Direction.left]
+  final Direction defaultDirection;
+
+  /// set controller for [DrawerScaffold]
+  final DrawerScaffoldController? controller;
+
+  /// set corner radius, default: 16.0
+  final double? cornerRadius;
+
+  /// set background color for [DrawerScaffold], default: [Theme.of(context).scaffoldBackgroundColor]
+  final Color? backgroundColor;
+
+  final bool extendedBody;
+  final bool? enableGestures;
+  final Widget? floatingActionButton;
+  final Widget? bottomNavigationBar;
+  final FloatingActionButtonLocation? floatingActionButtonLocation;
+  final FloatingActionButtonAnimator? floatingActionButtonAnimator;
+  final List<BoxShadow> contentShadow;
+  final Widget? bottomSheet;
+  final bool extendBodyBehindAppBar;
+  final List<Widget>? persistentFooterButtons;
+  final bool primary;
+  final bool? resizeToAvoidBottomInset;
+
+  /// Listen to offset value on slide event for which [SideDrawer]
+  final Function(SideDrawer, double)? onSlide;
+
+  /// Listen to which [SideDrawer] is opened (offset=1)
+  final Function(SideDrawer)? onOpened;
+
+  /// Listen to which [SideDrawer] is closed (offset=0)
+  final Function(SideDrawer)? onClosed;
 
   @override
   _DrawerScaffoldState createState() => new _DrawerScaffoldState();
 
   static MenuController currentController(BuildContext context,
       {bool nullOk = true}) {
-    assert(nullOk != null);
-    assert(context != null);
-    final _DrawerScaffoldState result =
+    final _DrawerScaffoldState? result =
         context.findAncestorStateOfType<_DrawerScaffoldState>();
-    if (nullOk || result != null) return result._controller;
+    if (nullOk || result != null) return result!._controller;
     throw FlutterError.fromParts(<DiagnosticsNode>[
       ErrorSummary(
           '_SideDrawerState.of() called with a context that does not contain a MenuController.'),
@@ -97,28 +104,28 @@ class DrawerScaffold extends StatefulWidget {
 
 class _DrawerScaffoldState<T> extends State<DrawerScaffold>
     with TickerProviderStateMixin {
-  List<MenuController> menuControllers;
+  List<MenuController>? menuControllers;
   // Curve scaleDownCurve = new Interval(0.0, 0.3, curve: Curves.easeOut);
   // Curve scaleUpCurve = new Interval(0.0, 1.0, curve: Curves.easeOut);
   // Curve slideOutCurve = new Interval(0.0, 1.0, curve: Curves.easeOut);
   // Curve slideInCurve = new Interval(0.0, 1.0, curve: Curves.easeOut);
 
-  Curve get scaleDownCurve => widget.drawers[focusDrawerIndex].scaleDownCurve;
-  Curve get scaleUpCurve => widget.drawers[focusDrawerIndex].scaleUpCurve;
-  Curve get slideOutCurve => widget.drawers[focusDrawerIndex].slideOutCurve;
-  Curve get slideInCurve => widget.drawers[focusDrawerIndex].slideInCurve;
+  Curve get scaleDownCurve => widget.drawers![focusDrawerIndex].scaleDownCurve;
+  Curve get scaleUpCurve => widget.drawers![focusDrawerIndex].scaleUpCurve;
+  Curve get slideOutCurve => widget.drawers![focusDrawerIndex].slideOutCurve;
+  Curve get slideInCurve => widget.drawers![focusDrawerIndex].slideInCurve;
 
   int listenDrawerIndex = 0;
   int focusDrawerIndex = 0;
 
   int get mainDrawerIndex => max(
       0,
-      widget.drawers
-          .indexWhere((element) => element.direction == widget.mainDrawer));
+      widget.drawers!.indexWhere(
+          (element) => element.direction == widget.defaultDirection));
   @override
   void initState() {
     super.initState();
-    selectedItemId = widget.drawers[listenDrawerIndex].selectedItemId;
+    selectedItemId = widget.drawers![listenDrawerIndex].selectedItemId;
 
     assignContoller();
 
@@ -138,7 +145,7 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
   @override
   void didUpdateWidget(Widget oldWidget) {
     assignContoller();
-    super.didUpdateWidget(oldWidget);
+    super.didUpdateWidget(oldWidget as DrawerScaffold);
   }
 
   MenuController createController(SideDrawer d) {
@@ -169,19 +176,19 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
 
   assignContoller() {
     if (menuControllers == null)
-      menuControllers ??= widget.drawers.map(createController).toList();
+      menuControllers ??= widget.drawers!.map(createController).toList();
     else
       for (var i = 0;
-          i < widget.drawers.length && i < menuControllers.length;
+          i < widget.drawers!.length && i < menuControllers!.length;
           i++) {
-        menuControllers[i]._drawer = widget.drawers[i];
+        menuControllers![i]._drawer = widget.drawers![i];
       }
-    for (var i = menuControllers.length; i < widget.drawers.length; i++) {
-      menuControllers.add(createController(widget.drawers[i]));
+    for (var i = menuControllers!.length; i < widget.drawers!.length; i++) {
+      menuControllers!.add(createController(widget.drawers![i]));
     }
     if (widget.controller != null) {
-      widget.controller._menuControllers = menuControllers;
-      widget.controller._setFocus = (index) {
+      widget.controller!._menuControllers = menuControllers;
+      widget.controller!._setFocus = (index) {
         focusDrawerIndex = index;
       };
     }
@@ -189,47 +196,47 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
 
   void updateDrawerState() {
     if (widget.controller != null) {
-      if (widget.controller._open != null)
-        menuControllers
+      if (widget.controller!._open != null)
+        menuControllers!
             .firstWhere((element) =>
-                element._drawer.direction == widget.controller._open)
+                element._drawer.direction == widget.controller!._open)
             .open();
       else
-        menuControllers.forEach((element) {
+        menuControllers!.forEach((element) {
           element.close();
         });
     }
   }
 
-  Widget createAppBar() {
+  Widget? createAppBar() {
     if (widget.appBar != null)
       return AppBar(
-          actionsIconTheme: widget.appBar.actionsIconTheme,
-          excludeHeaderSemantics: widget.appBar.excludeHeaderSemantics,
-          shape: widget.appBar.shape,
-          key: widget.appBar.key,
-          backgroundColor: widget.appBar.backgroundColor,
-          leading: widget.appBar.leading ??
+          actionsIconTheme: widget.appBar!.actionsIconTheme,
+          excludeHeaderSemantics: widget.appBar!.excludeHeaderSemantics,
+          shape: widget.appBar!.shape,
+          key: widget.appBar!.key,
+          backgroundColor: widget.appBar!.backgroundColor,
+          leading: widget.appBar!.leading ??
               new IconButton(
                   icon: Icon(Icons.menu),
                   onPressed: () {
                     focusDrawerIndex = mainDrawerIndex;
-                    menuControllers[mainDrawerIndex].toggle();
+                    menuControllers![mainDrawerIndex].toggle();
                   }),
-          title: widget.appBar.title,
-          automaticallyImplyLeading: widget.appBar.automaticallyImplyLeading,
-          actions: widget.appBar.actions,
-          flexibleSpace: widget.appBar.flexibleSpace,
-          bottom: widget.appBar.bottom,
-          elevation: widget.appBar.elevation,
-          brightness: widget.appBar.brightness,
-          iconTheme: widget.appBar.iconTheme,
-          textTheme: widget.appBar.textTheme,
-          primary: widget.appBar.primary,
-          centerTitle: widget.appBar.centerTitle,
-          titleSpacing: widget.appBar.titleSpacing,
-          toolbarOpacity: widget.appBar.toolbarOpacity,
-          bottomOpacity: widget.appBar.bottomOpacity);
+          title: widget.appBar!.title,
+          automaticallyImplyLeading: widget.appBar!.automaticallyImplyLeading,
+          actions: widget.appBar!.actions,
+          flexibleSpace: widget.appBar!.flexibleSpace,
+          bottom: widget.appBar!.bottom,
+          elevation: widget.appBar!.elevation,
+          brightness: widget.appBar!.brightness,
+          iconTheme: widget.appBar!.iconTheme,
+          textTheme: widget.appBar!.textTheme,
+          primary: widget.appBar!.primary,
+          centerTitle: widget.appBar!.centerTitle,
+          titleSpacing: widget.appBar!.titleSpacing,
+          toolbarOpacity: widget.appBar!.toolbarOpacity,
+          bottomOpacity: widget.appBar!.bottomOpacity);
 
     return null;
   }
@@ -238,30 +245,28 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
   double percentage = 0.0;
   bool isOpening = false;
 
-  Widget body;
+  Widget? body;
 
-  T selectedItemId;
+  T? selectedItemId;
   bool isDrawerOpen() {
-    return menuControllers.where((element) => element.isOpen()).isNotEmpty;
+    return menuControllers!.where((element) => element.isOpen()).isNotEmpty;
   }
 
   int drawerFrom(Direction direction) {
-    return menuControllers.indexWhere((element) {
+    return menuControllers!.indexWhere((element) {
       return element._drawer.direction == direction;
     });
   }
 
   createContentDisplay() {
-    if (selectedItemId != widget.drawers[listenDrawerIndex].selectedItemId ||
+    if (selectedItemId != widget.drawers![listenDrawerIndex].selectedItemId ||
         body == null) {
-      selectedItemId = widget.drawers[listenDrawerIndex].selectedItemId;
-      body = widget.builder?.call(context, selectedItemId) ??
-          // ignore: deprecated_member_use_from_same_package
-          widget.contentView?.contentBuilder(context);
+      selectedItemId = widget.drawers![listenDrawerIndex].selectedItemId;
+      body = widget.builder?.call(context, selectedItemId);
     }
     Widget _scaffoldWidget = new Scaffold(
       backgroundColor: Colors.transparent,
-      appBar: createAppBar(),
+      appBar: createAppBar() as PreferredSizeWidget?,
       body: body,
       extendBody: widget.extendedBody,
       floatingActionButton: widget.floatingActionButton,
@@ -273,19 +278,18 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
       persistentFooterButtons: widget.persistentFooterButtons,
       primary: widget.primary,
       resizeToAvoidBottomInset: widget.resizeToAvoidBottomInset,
-      resizeToAvoidBottomPadding: widget.resizeToAvoidBottomPadding,
     );
 
     double maxSlideAmount =
-        widget.drawers[focusDrawerIndex].maxSlideAmount(context);
-    Widget content = !widget.enableGestures
+        widget.drawers![focusDrawerIndex].maxSlideAmount(context);
+    Widget content = !widget.enableGestures!
         ? _scaffoldWidget
         : GestureDetector(
             child: AbsorbPointer(
                 absorbing: isDrawerOpen() && widget.appBar != null,
                 child: _scaffoldWidget),
             onTap: () {
-              menuControllers.forEach((element) {
+              menuControllers!.forEach((element) {
                 if (element.isOpen()) element.close();
               });
             },
@@ -326,7 +330,8 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
               if (startDx == -1) return;
 
               double dx = (details.globalPosition.dx - startDx);
-              MenuController menuController = menuControllers[focusDrawerIndex];
+              MenuController menuController =
+                  menuControllers![focusDrawerIndex];
               if (menuController._drawer.direction == Direction.right) {
                 dx = -dx;
               }
@@ -348,7 +353,7 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
             },
             onHorizontalDragEnd: (details) {
               if (startDx == -1) return;
-              menuControllers.forEach((menuController) {
+              menuControllers!.forEach((menuController) {
                 if (percentage < 0.5) {
                   menuController.close();
                 } else {
@@ -362,10 +367,8 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
 
     return zoomAndSlideContent(new Container(
         decoration: new BoxDecoration(
-          // ignore: deprecated_member_use_from_same_package
-          image: widget.contentView?.background,
-          // ignore: deprecated_member_use_from_same_package
-          color: widget.contentView?.color ?? Theme.of(context).canvasColor,
+          color: widget.backgroundColor ??
+              Theme.of(context).scaffoldBackgroundColor,
         ),
         child: isIOS
             ? content
@@ -374,7 +377,7 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
                 onWillPop: () {
                   return Future(() {
                     if (isDrawerOpen()) {
-                      menuControllers.forEach((element) {
+                      menuControllers!.forEach((element) {
                         element.close();
                       });
                       return false;
@@ -385,12 +388,12 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
   }
 
   zoomAndSlideContent(Widget content) {
-    SideDrawer drawer = widget.drawers[focusDrawerIndex];
-    MenuController menuController = this.menuControllers[focusDrawerIndex];
+    SideDrawer drawer = widget.drawers![focusDrawerIndex];
+    MenuController menuController = this.menuControllers![focusDrawerIndex];
     double slidePercent = menuController._slidePercent;
     double contentScale = menuController.contentScale;
     double slideAmount = menuController.slideAmount;
-    final cornerRadius = (drawer.cornerRadius ?? widget.cornerRadius) *
+    final cornerRadius = (drawer.cornerRadius ?? widget.cornerRadius)! *
         menuController.percentOpen;
 
     double degreeAmount = (drawer.degree ?? 0) * slidePercent;
@@ -438,22 +441,21 @@ class _DrawerScaffoldState<T> extends State<DrawerScaffold>
 
   @override
   Widget build(BuildContext context) {
-    focusDrawerIndex = min(
-        widget.drawers.length - 1, focusDrawerIndex ?? widget.drawers.length);
+    focusDrawerIndex = min(widget.drawers!.length - 1, focusDrawerIndex);
     return Stack(
       children: [
-        focusDrawerIndex >= 0 ? widget.drawers[focusDrawerIndex] : Container(),
+        focusDrawerIndex >= 0 ? widget.drawers![focusDrawerIndex] : Container(),
         createContentDisplay(),
       ],
     );
   }
 
-  MenuController get _controller => menuControllers[focusDrawerIndex];
+  MenuController get _controller => menuControllers![focusDrawerIndex];
 }
 
 class DrawerScaffoldMenuController extends StatefulWidget {
-  final DrawerScaffoldBuilder builder;
-  final Direction direction;
+  final DrawerScaffoldBuilder? builder;
+  final Direction? direction;
   DrawerScaffoldMenuController({
     this.builder,
     this.direction,
@@ -467,38 +469,38 @@ class DrawerScaffoldMenuController extends StatefulWidget {
 
 class DrawerScaffoldMenuControllerState
     extends State<DrawerScaffoldMenuController> {
-  MenuController menuController;
+  MenuController? menuController;
 
   @override
   void initState() {
     super.initState();
 
     menuController = getMenuController(context, widget.direction);
-    menuController.addListener(_onMenuControllerChange);
+    menuController!.addListener(_onMenuControllerChange);
   }
 
   @override
   void didUpdateWidget(Widget oldWidget) {
-    super.didUpdateWidget(oldWidget);
+    super.didUpdateWidget(oldWidget as DrawerScaffoldMenuController);
     if (menuController != null)
-      menuController.removeListener(_onMenuControllerChange);
+      menuController!.removeListener(_onMenuControllerChange);
     menuController = getMenuController(context, widget.direction);
-    menuController.addListener(_onMenuControllerChange);
+    menuController!.addListener(_onMenuControllerChange);
   }
 
   @override
   void dispose() {
-    menuController.removeListener(_onMenuControllerChange);
+    menuController!.removeListener(_onMenuControllerChange);
     super.dispose();
   }
 
   MenuController getMenuController(BuildContext context,
-      [Direction direction = Direction.left]) {
+      [Direction? direction = Direction.left]) {
     final scaffoldState =
-        context.findAncestorStateOfType<_DrawerScaffoldState>();
-    return scaffoldState.menuControllers.firstWhere(
+        context.findAncestorStateOfType<_DrawerScaffoldState>()!;
+    return scaffoldState.menuControllers!.firstWhere(
       (element) => element._drawer.direction == direction,
-      orElse: () => scaffoldState.menuControllers[0],
+      orElse: () => scaffoldState.menuControllers![0],
     );
   }
 
@@ -508,21 +510,21 @@ class DrawerScaffoldMenuControllerState
 
   @override
   Widget build(BuildContext context) {
-    return widget.builder(
+    return widget.builder!(
         context, getMenuController(context, widget.direction));
   }
 }
 
-typedef Widget ScreenBuilder<T>(BuildContext context, T id);
+typedef Widget ScreenBuilder<T>(BuildContext context, T? id);
 
 class Screen {
-  final String title;
-  final DecorationImage background;
-  final WidgetBuilder contentBuilder;
+  final String? title;
+  final DecorationImage? background;
+  final WidgetBuilder? contentBuilder;
 
-  final Color color;
+  final Color? color;
 
-  final Color appBarColor;
+  final Color? appBarColor;
 
   final bool enableGestures;
 
@@ -549,11 +551,11 @@ class MenuController extends ChangeNotifier {
   double get slidePercent => _slidePercent;
   double get scalePercent => _scalePercent;
   MenuController(this._drawer, this.onAnimated,
-      {this.vsync, BuildContext context})
+      {required this.vsync, BuildContext? context})
       : this.duration = _drawer.duration ?? const Duration(milliseconds: 250),
         _animationController = new AnimationController(vsync: vsync) {
     _animationController
-      ..duration = duration ?? const Duration(milliseconds: 250)
+      ..duration = duration
       ..addListener(() {
         calculate(context);
 
@@ -582,7 +584,7 @@ class MenuController extends ChangeNotifier {
     calculate(context);
   }
 
-  calculate(BuildContext context) {
+  calculate(BuildContext? context) {
     switch (state) {
       case MenuState.closed:
         _slidePercent = 0.0;
@@ -643,13 +645,13 @@ class MenuController extends ChangeNotifier {
 }
 
 class DrawerScaffoldController {
-  List<MenuController> _menuControllers;
+  List<MenuController>? _menuControllers;
 
-  ValueChanged<int> _setFocus;
+  late ValueChanged<int> _setFocus;
 
-  DrawerScaffoldController({Direction open}) : _open = open;
+  DrawerScaffoldController({Direction? open}) : _open = open;
 
-  Direction _open;
+  Direction? _open;
   toggle([Direction direction = Direction.left]) {
     if (isOpen())
       closeDrawer(direction);
@@ -658,23 +660,23 @@ class DrawerScaffoldController {
   }
 
   openDrawer([Direction direction = Direction.left]) {
-    int index = _menuControllers
+    int index = _menuControllers!
         .indexWhere((element) => element._drawer.direction == direction);
     if (index >= 0) {
       _setFocus(index);
-      _menuControllers[index].open();
+      _menuControllers![index].open();
     }
   }
 
   closeDrawer([Direction direction = Direction.left]) {
-    _menuControllers
+    _menuControllers!
         .firstWhere((element) => element._drawer.direction == direction)
         .close();
   }
 
-  ValueChanged<bool> onToggle;
+  ValueChanged<bool>? onToggle;
 
-  bool isOpen([Direction direction = Direction.left]) => _menuControllers
+  bool isOpen([Direction direction = Direction.left]) => _menuControllers!
       .where((element) =>
           element._drawer.direction == direction && element.isOpen())
       .isNotEmpty;